refactor: split tools_quality and evidence tests into sibling test files
Four review rounds of in-file regression tests pushed both modules past
the R8 1400-LOC hard cap. Converted to the directory-module pattern
ai_generations and validate/content already use (mod.rs + cfg(test)
test.rs); pure code movement, test counts unchanged (33 + 24).
